# test: bgworker parallel database