--- a/Cargo.toml
+++ b/Cargo.toml
@@ -210,4 +210,5 @@ sqlx = { version = "0.8.2", default-features = false, features = [
     "postgres",
     "chrono",
     "sqlite",
+    "migrate",
 ] }